Split RGU test symbols into separate files
